requirements.txt
pillow
aiogram
python-dotenv
